fix(drawsystem): delete result package and SingleDrawResult
delete result package and SingleDrawResult class due bc they are useless
