+++ src/main/java/lotto/drawsystem/DrawResults.java
@@ -1,0 +1,14 @@
+package lotto.drawsystem;
+
+import java.util.List;
+
+public class DrawResults {
+
+    List<List<Integer>> totalMainNumbers;
+    List<Integer> totalBonusNumbers;
+
+    public DrawResults(List<List<Integer>> mainNumbers, List<Integer> totalBonusNumbers) {
+        this.totalMainNumbers = mainNumbers;
+        this.totalBonusNumbers = totalBonusNumbers;
+    }
+}
